update generate-schema script to generate files in airtable folder
