+++ src/lib/airtable/request.js
@@ -1,6 +1,7 @@
+/* eslint-disable */ 
+/* eslint no-restricted-imports: 0 */
 /* eslint no-restricted-imports: 0 */
 /* eslint-disable no-unused-vars */
-/* eslint-disable */ 
 
 /*
   THIS IS A GENERATED FILE
@@ -59,21 +60,6 @@ export const createManyFarms = async (records) => {
   return Promise.all(createPromises);
 };
 
-export const createCluster = async (record) => {
-  return createRecord(Tables.Clusters, record);
-};
-
-export const createManyClusters = async (records) => {
-  const createPromises = [];
-  const numCalls = Math.ceil(records.length / 10);
-  for (let i = 0; i < numCalls; i += 1) {
-    const subset = records.slice(i * 10, (i + 1) * 10);
-    if (subset.length > 0)
-      createPromises.push(createRecords(Tables.Clusters, subset));
-  }
-  return Promise.all(createPromises);
-};
-
 export const createAgencie = async (record) => {
   return createRecord(Tables.Agencies, record);
 };
@@ -85,21 +71,6 @@ export const createManyAgencies = async (records) => {
     const subset = records.slice(i * 10, (i + 1) * 10);
     if (subset.length > 0)
       createPromises.push(createRecords(Tables.Agencies, subset));
-  }
-  return Promise.all(createPromises);
-};
-
-export const createListing = async (record) => {
-  return createRecord(Tables.Listings, record);
-};
-
-export const createManyListings = async (records) => {
-  const createPromises = [];
-  const numCalls = Math.ceil(records.length / 10);
-  for (let i = 0; i < numCalls; i += 1) {
-    const subset = records.slice(i * 10, (i + 1) * 10);
-    if (subset.length > 0)
-      createPromises.push(createRecords(Tables.Listings, subset));
   }
   return Promise.all(createPromises);
 };
@@ -119,17 +90,17 @@ export const createManyQuotes = async (records) => {
   return Promise.all(createPromises);
 };
 
-export const createReservedListing = async (record) => {
-  return createRecord(Tables.ReservedListings, record);
+export const createCheckoutProfile = async (record) => {
+  return createRecord(Tables.CheckoutProfiles, record);
 };
 
-export const createManyReservedListings = async (records) => {
+export const createManyCheckoutProfiles = async (records) => {
   const createPromises = [];
   const numCalls = Math.ceil(records.length / 10);
   for (let i = 0; i < numCalls; i += 1) {
     const subset = records.slice(i * 10, (i + 1) * 10);
     if (subset.length > 0)
-      createPromises.push(createRecords(Tables.ReservedListings, subset));
+      createPromises.push(createRecords(Tables.CheckoutProfiles, subset));
   }
   return Promise.all(createPromises);
 };
@@ -149,6 +120,81 @@ export const createManyNotifications = async (records) => {
   return Promise.all(createPromises);
 };
 
+export const createEmail = async (record) => {
+  return createRecord(Tables.Email, record);
+};
+
+export const createManyEmails = async (records) => {
+  const createPromises = [];
+  const numCalls = Math.ceil(records.length / 10);
+  for (let i = 0; i < numCalls; i += 1) {
+    const subset = records.slice(i * 10, (i + 1) * 10);
+    if (subset.length > 0)
+      createPromises.push(createRecords(Tables.Email, subset));
+  }
+  return Promise.all(createPromises);
+};
+
+export const createReservedListing = async (record) => {
+  return createRecord(Tables.ReservedListings, record);
+};
+
+export const createManyReservedListings = async (records) => {
+  const createPromises = [];
+  const numCalls = Math.ceil(records.length / 10);
+  for (let i = 0; i < numCalls; i += 1) {
+    const subset = records.slice(i * 10, (i + 1) * 10);
+    if (subset.length > 0)
+      createPromises.push(createRecords(Tables.ReservedListings, subset));
+  }
+  return Promise.all(createPromises);
+};
+
+export const createListing = async (record) => {
+  return createRecord(Tables.Listings, record);
+};
+
+export const createManyListings = async (records) => {
+  const createPromises = [];
+  const numCalls = Math.ceil(records.length / 10);
+  for (let i = 0; i < numCalls; i += 1) {
+    const subset = records.slice(i * 10, (i + 1) * 10);
+    if (subset.length > 0)
+      createPromises.push(createRecords(Tables.Listings, subset));
+  }
+  return Promise.all(createPromises);
+};
+
+export const createListingUPDATED = async (record) => {
+  return createRecord(Tables.ListingUPDATED, record);
+};
+
+export const createManyListingUPDATEDs = async (records) => {
+  const createPromises = [];
+  const numCalls = Math.ceil(records.length / 10);
+  for (let i = 0; i < numCalls; i += 1) {
+    const subset = records.slice(i * 10, (i + 1) * 10);
+    if (subset.length > 0)
+      createPromises.push(createRecords(Tables.ListingUPDATED, subset));
+  }
+  return Promise.all(createPromises);
+};
+
+export const createProduceType = async (record) => {
+  return createRecord(Tables.ProduceType, record);
+};
+
+export const createManyProduceTypes = async (records) => {
+  const createPromises = [];
+  const numCalls = Math.ceil(records.length / 10);
+  for (let i = 0; i < numCalls; i += 1) {
+    const subset = records.slice(i * 10, (i + 1) * 10);
+    if (subset.length > 0)
+      createPromises.push(createRecords(Tables.ProduceType, subset));
+  }
+  return Promise.all(createPromises);
+};
+
 /*
  ******* READ RECORDS *******
  */
@@ -160,7 +206,7 @@ export const getUserById = async (id) => {
 export const getUsersByIds = async ( ids, filterByFormula = '', sort = []
 ) => {
   let formula = `OR(${ids.reduce(
-    (f, id) => `${f} {ID}='${id}',`,
+    (f, id) => `${f} RECORD_ID()='${id}',`,
     ''
   )} 1 < 0)`;
   formula = filterByFormula ? `AND(${filterByFormula}, ${formula})` : formula;
@@ -178,7 +224,7 @@ export const getFarmById = async (id) => {
 export const getFarmsByIds = async ( ids, filterByFormula = '', sort = []
 ) => {
   let formula = `OR(${ids.reduce(
-    (f, id) => `${f} {ID}='${id}',`,
+    (f, id) => `${f} RECORD_ID()='${id}',`,
     ''
   )} 1 < 0)`;
   formula = filterByFormula ? `AND(${filterByFormula}, ${formula})` : formula;
@@ -189,24 +235,6 @@ export const getAllFarms = async (filterByFormula = '', sort = []) => {
   return getAllRecords(Tables.Farms, filterByFormula, sort);
 };
 
-export const getClusterById = async (id) => {
-  return getRecordById(Tables.Clusters, id);
-};
-
-export const getClustersByIds = async ( ids, filterByFormula = '', sort = []
-) => {
-  let formula = `OR(${ids.reduce(
-    (f, id) => `${f} {ID}='${id}',`,
-    ''
-  )} 1 < 0)`;
-  formula = filterByFormula ? `AND(${filterByFormula}, ${formula})` : formula;
-  return getAllRecords(Tables.Clusters, formula, sort);
-};
-
-export const getAllClusters = async (filterByFormula = '', sort = []) => {
-  return getAllRecords(Tables.Clusters, filterByFormula, sort);
-};
-
 export const getAgencieById = async (id) => {
   return getRecordById(Tables.Agencies, id);
 };
@@ -214,7 +242,7 @@ export const getAgencieById = async (id) => {
 export const getAgenciesByIds = async ( ids, filterByFormula = '', sort = []
 ) => {
   let formula = `OR(${ids.reduce(
-    (f, id) => `${f} {ID}='${id}',`,
+    (f, id) => `${f} RECORD_ID()='${id}',`,
     ''
   )} 1 < 0)`;
   formula = filterByFormula ? `AND(${filterByFormula}, ${formula})` : formula;
@@ -225,24 +253,6 @@ export const getAllAgencies = async (filterByFormula = '', sort = []) => {
   return getAllRecords(Tables.Agencies, filterByFormula, sort);
 };
 
-export const getListingById = async (id) => {
-  return getRecordById(Tables.Listings, id);
-};
-
-export const getListingsByIds = async ( ids, filterByFormula = '', sort = []
-) => {
-  let formula = `OR(${ids.reduce(
-    (f, id) => `${f} {ID}='${id}',`,
-    ''
-  )} 1 < 0)`;
-  formula = filterByFormula ? `AND(${filterByFormula}, ${formula})` : formula;
-  return getAllRecords(Tables.Listings, formula, sort);
-};
-
-export const getAllListings = async (filterByFormula = '', sort = []) => {
-  return getAllRecords(Tables.Listings, filterByFormula, sort);
-};
-
 export const getQuoteById = async (id) => {
   return getRecordById(Tables.Quotes, id);
 };
@@ -250,7 +260,7 @@ export const getQuoteById = async (id) => {
 export const getQuotesByIds = async ( ids, filterByFormula = '', sort = []
 ) => {
   let formula = `OR(${ids.reduce(
-    (f, id) => `${f} {ID}='${id}',`,
+    (f, id) => `${f} RECORD_ID()='${id}',`,
     ''
   )} 1 < 0)`;
   formula = filterByFormula ? `AND(${filterByFormula}, ${formula})` : formula;
@@ -261,6 +271,60 @@ export const getAllQuotes = async (filterByFormula = '', sort = []) => {
   return getAllRecords(Tables.Quotes, filterByFormula, sort);
 };
 
+export const getCheckoutProfileById = async (id) => {
+  return getRecordById(Tables.CheckoutProfiles, id);
+};
+
+export const getCheckoutProfilesByIds = async ( ids, filterByFormula = '', sort = []
+) => {
+  let formula = `OR(${ids.reduce(
+    (f, id) => `${f} RECORD_ID()='${id}',`,
+    ''
+  )} 1 < 0)`;
+  formula = filterByFormula ? `AND(${filterByFormula}, ${formula})` : formula;
+  return getAllRecords(Tables.CheckoutProfiles, formula, sort);
+};
+
+export const getAllCheckoutProfiles = async (filterByFormula = '', sort = []) => {
+  return getAllRecords(Tables.CheckoutProfiles, filterByFormula, sort);
+};
+
+export const getNotificationById = async (id) => {
+  return getRecordById(Tables.Notifications, id);
+};
+
+export const getNotificationsByIds = async ( ids, filterByFormula = '', sort = []
+) => {
+  let formula = `OR(${ids.reduce(
+    (f, id) => `${f} RECORD_ID()='${id}',`,
+    ''
+  )} 1 < 0)`;
+  formula = filterByFormula ? `AND(${filterByFormula}, ${formula})` : formula;
+  return getAllRecords(Tables.Notifications, formula, sort);
+};
+
+export const getAllNotifications = async (filterByFormula = '', sort = []) => {
+  return getAllRecords(Tables.Notifications, filterByFormula, sort);
+};
+
+export const getEmailById = async (id) => {
+  return getRecordById(Tables.Email, id);
+};
+
+export const getEmailsByIds = async ( ids, filterByFormula = '', sort = []
+) => {
+  let formula = `OR(${ids.reduce(
+    (f, id) => `${f} RECORD_ID()='${id}',`,
+    ''
+  )} 1 < 0)`;
+  formula = filterByFormula ? `AND(${filterByFormula}, ${formula})` : formula;
+  return getAllRecords(Tables.Email, formula, sort);
+};
+
+export const getAllEmails = async (filterByFormula = '', sort = []) => {
+  return getAllRecords(Tables.Email, filterByFormula, sort);
+};
+
 export const getReservedListingById = async (id) => {
   return getRecordById(Tables.ReservedListings, id);
 };
@@ -268,7 +332,7 @@ export const getReservedListingById = async (id) => {
 export const getReservedListingsByIds = async ( ids, filterByFormula = '', sort = []
 ) => {
   let formula = `OR(${ids.reduce(
-    (f, id) => `${f} {ID}='${id}',`,
+    (f, id) => `${f} RECORD_ID()='${id}',`,
     ''
   )} 1 < 0)`;
   formula = filterByFormula ? `AND(${filterByFormula}, ${formula})` : formula;
@@ -279,22 +343,58 @@ export const getAllReservedListings = async (filterByFormula = '', sort = []) =>
   return getAllRecords(Tables.ReservedListings, filterByFormula, sort);
 };
 
-export const getNotificationById = async (id) => {
-  return getRecordById(Tables.Notifications, id);
+export const getListingById = async (id) => {
+  return getRecordById(Tables.Listings, id);
 };
 
-export const getNotificationsByIds = async ( ids, filterByFormula = '', sort = []
+export const getListingsByIds = async ( ids, filterByFormula = '', sort = []
 ) => {
   let formula = `OR(${ids.reduce(
-    (f, id) => `${f} {ID}='${id}',`,
+    (f, id) => `${f} RECORD_ID()='${id}',`,
     ''
   )} 1 < 0)`;
   formula = filterByFormula ? `AND(${filterByFormula}, ${formula})` : formula;
-  return getAllRecords(Tables.Notifications, formula, sort);
+  return getAllRecords(Tables.Listings, formula, sort);
 };
 
-export const getAllNotifications = async (filterByFormula = '', sort = []) => {
-  return getAllRecords(Tables.Notifications, filterByFormula, sort);
+export const getAllListings = async (filterByFormula = '', sort = []) => {
+  return getAllRecords(Tables.Listings, filterByFormula, sort);
+};
+
+export const getListingUPDATEDById = async (id) => {
+  return getRecordById(Tables.ListingUPDATED, id);
+};
+
+export const getListingUPDATEDsByIds = async ( ids, filterByFormula = '', sort = []
+) => {
+  let formula = `OR(${ids.reduce(
+    (f, id) => `${f} RECORD_ID()='${id}',`,
+    ''
+  )} 1 < 0)`;
+  formula = filterByFormula ? `AND(${filterByFormula}, ${formula})` : formula;
+  return getAllRecords(Tables.ListingUPDATED, formula, sort);
+};
+
+export const getAllListingUPDATEDs = async (filterByFormula = '', sort = []) => {
+  return getAllRecords(Tables.ListingUPDATED, filterByFormula, sort);
+};
+
+export const getProduceTypeById = async (id) => {
+  return getRecordById(Tables.ProduceType, id);
+};
+
+export const getProduceTypesByIds = async ( ids, filterByFormula = '', sort = []
+) => {
+  let formula = `OR(${ids.reduce(
+    (f, id) => `${f} RECORD_ID()='${id}',`,
+    ''
+  )} 1 < 0)`;
+  formula = filterByFormula ? `AND(${filterByFormula}, ${formula})` : formula;
+  return getAllRecords(Tables.ProduceType, formula, sort);
+};
+
+export const getAllProduceTypes = async (filterByFormula = '', sort = []) => {
+  return getAllRecords(Tables.ProduceType, filterByFormula, sort);
 };
 
 /*
@@ -331,21 +431,6 @@ export const updateManyFarms = async (recordUpdates) => {
   return Promise.all(updatePromises);
 };
 
-export const updateCluster = async (id, recordUpdates) => {
-  return updateRecord(Tables.Clusters, id, recordUpdates);
-};
-
-export const updateManyClusters = async (recordUpdates) => {
-  const updatePromises = [];
-  const numCalls = Math.ceil(recordUpdates.length / 10);
-  for (let i = 0; i < numCalls; i += 1) {
-    const subset = recordUpdates.slice(i * 10, (i + 1) * 10);
-    if (subset.length > 0)
-      updatePromises.push(updateRecords(Tables.Clusters, subset));
-  }
-  return Promise.all(updatePromises);
-};
-
 export const updateAgencie = async (id, recordUpdates) => {
   return updateRecord(Tables.Agencies, id, recordUpdates);
 };
@@ -357,21 +442,6 @@ export const updateManyAgencies = async (recordUpdates) => {
     const subset = recordUpdates.slice(i * 10, (i + 1) * 10);
     if (subset.length > 0)
       updatePromises.push(updateRecords(Tables.Agencies, subset));
-  }
-  return Promise.all(updatePromises);
-};
-
-export const updateListing = async (id, recordUpdates) => {
-  return updateRecord(Tables.Listings, id, recordUpdates);
-};
-
-export const updateManyListings = async (recordUpdates) => {
-  const updatePromises = [];
-  const numCalls = Math.ceil(recordUpdates.length / 10);
-  for (let i = 0; i < numCalls; i += 1) {
-    const subset = recordUpdates.slice(i * 10, (i + 1) * 10);
-    if (subset.length > 0)
-      updatePromises.push(updateRecords(Tables.Listings, subset));
   }
   return Promise.all(updatePromises);
 };
@@ -391,17 +461,17 @@ export const updateManyQuotes = async (recordUpdates) => {
   return Promise.all(updatePromises);
 };
 
-export const updateReservedListing = async (id, recordUpdates) => {
-  return updateRecord(Tables.ReservedListings, id, recordUpdates);
+export const updateCheckoutProfile = async (id, recordUpdates) => {
+  return updateRecord(Tables.CheckoutProfiles, id, recordUpdates);
 };
 
-export const updateManyReservedListings = async (recordUpdates) => {
+export const updateManyCheckoutProfiles = async (recordUpdates) => {
   const updatePromises = [];
   const numCalls = Math.ceil(recordUpdates.length / 10);
   for (let i = 0; i < numCalls; i += 1) {
     const subset = recordUpdates.slice(i * 10, (i + 1) * 10);
     if (subset.length > 0)
-      updatePromises.push(updateRecords(Tables.ReservedListings, subset));
+      updatePromises.push(updateRecords(Tables.CheckoutProfiles, subset));
   }
   return Promise.all(updatePromises);
 };
@@ -421,6 +491,81 @@ export const updateManyNotifications = async (recordUpdates) => {
   return Promise.all(updatePromises);
 };
 
+export const updateEmail = async (id, recordUpdates) => {
+  return updateRecord(Tables.Email, id, recordUpdates);
+};
+
+export const updateManyEmails = async (recordUpdates) => {
+  const updatePromises = [];
+  const numCalls = Math.ceil(recordUpdates.length / 10);
+  for (let i = 0; i < numCalls; i += 1) {
+    const subset = recordUpdates.slice(i * 10, (i + 1) * 10);
+    if (subset.length > 0)
+      updatePromises.push(updateRecords(Tables.Email, subset));
+  }
+  return Promise.all(updatePromises);
+};
+
+export const updateReservedListing = async (id, recordUpdates) => {
+  return updateRecord(Tables.ReservedListings, id, recordUpdates);
+};
+
+export const updateManyReservedListings = async (recordUpdates) => {
+  const updatePromises = [];
+  const numCalls = Math.ceil(recordUpdates.length / 10);
+  for (let i = 0; i < numCalls; i += 1) {
+    const subset = recordUpdates.slice(i * 10, (i + 1) * 10);
+    if (subset.length > 0)
+      updatePromises.push(updateRecords(Tables.ReservedListings, subset));
+  }
+  return Promise.all(updatePromises);
+};
+
+export const updateListing = async (id, recordUpdates) => {
+  return updateRecord(Tables.Listings, id, recordUpdates);
+};
+
+export const updateManyListings = async (recordUpdates) => {
+  const updatePromises = [];
+  const numCalls = Math.ceil(recordUpdates.length / 10);
+  for (let i = 0; i < numCalls; i += 1) {
+    const subset = recordUpdates.slice(i * 10, (i + 1) * 10);
+    if (subset.length > 0)
+      updatePromises.push(updateRecords(Tables.Listings, subset));
+  }
+  return Promise.all(updatePromises);
+};
+
+export const updateListingUPDATED = async (id, recordUpdates) => {
+  return updateRecord(Tables.ListingUPDATED, id, recordUpdates);
+};
+
+export const updateManyListingUPDATEDs = async (recordUpdates) => {
+  const updatePromises = [];
+  const numCalls = Math.ceil(recordUpdates.length / 10);
+  for (let i = 0; i < numCalls; i += 1) {
+    const subset = recordUpdates.slice(i * 10, (i + 1) * 10);
+    if (subset.length > 0)
+      updatePromises.push(updateRecords(Tables.ListingUPDATED, subset));
+  }
+  return Promise.all(updatePromises);
+};
+
+export const updateProduceType = async (id, recordUpdates) => {
+  return updateRecord(Tables.ProduceType, id, recordUpdates);
+};
+
+export const updateManyProduceTypes = async (recordUpdates) => {
+  const updatePromises = [];
+  const numCalls = Math.ceil(recordUpdates.length / 10);
+  for (let i = 0; i < numCalls; i += 1) {
+    const subset = recordUpdates.slice(i * 10, (i + 1) * 10);
+    if (subset.length > 0)
+      updatePromises.push(updateRecords(Tables.ProduceType, subset));
+  }
+  return Promise.all(updatePromises);
+};
+
 /*
  ******* DELETE RECORDS *******
  */
@@ -431,21 +576,30 @@ export const deleteUser = async (id) => {
 export const deleteFarm = async (id) => {
   return deleteRecord(Tables.Farms, id);
 };
-export const deleteCluster = async (id) => {
-  return deleteRecord(Tables.Clusters, id);
-};
 export const deleteAgencie = async (id) => {
   return deleteRecord(Tables.Agencies, id);
-};
-export const deleteListing = async (id) => {
-  return deleteRecord(Tables.Listings, id);
 };
 export const deleteQuote = async (id) => {
   return deleteRecord(Tables.Quotes, id);
 };
-export const deleteReservedListing = async (id) => {
-  return deleteRecord(Tables.ReservedListings, id);
+export const deleteCheckoutProfile = async (id) => {
+  return deleteRecord(Tables.CheckoutProfiles, id);
 };
 export const deleteNotification = async (id) => {
   return deleteRecord(Tables.Notifications, id);
+};
+export const deleteEmail = async (id) => {
+  return deleteRecord(Tables.Email, id);
+};
+export const deleteReservedListing = async (id) => {
+  return deleteRecord(Tables.ReservedListings, id);
+};
+export const deleteListing = async (id) => {
+  return deleteRecord(Tables.Listings, id);
+};
+export const deleteListingUPDATED = async (id) => {
+  return deleteRecord(Tables.ListingUPDATED, id);
+};
+export const deleteProduceType = async (id) => {
+  return deleteRecord(Tables.ProduceType, id);
 };
