+++ src/lib/airtable/schema.js
@@ -1,18 +1,23 @@
+/* eslint-disable */ 
+/* eslint no-restricted-imports: 0 */
+/* eslint no-restricted-imports: 0 */
 /*
     THIS IS A GENERATED FILE
     Changes might be overwritten in the future, edit with caution!
 */
-/* eslint-disable */ 
 
 export const Tables = {
 	Users: 'Users',
 	Farms: 'Farms',
-	Clusters: 'Clusters',
 	Agencies: 'Agencies',
-	Listings: 'Listings',
 	Quotes: 'Quotes',
-	ReservedListings: 'Reserved Listings',
+	CheckoutProfiles: 'Checkout Profiles',
 	Notifications: 'Notifications',
+	Email: 'Email',
+	ReservedListings: 'Reserved Listings',
+	Listings: 'Listings',
+	ListingUPDATED: 'Listing UPDATED',
+	ProduceType: 'Produce Type',
 };
 
 export const Columns = {
@@ -20,13 +25,20 @@ export const Columns = {
 		username: {name:`username`, type:`text`},
 		password: {name:`password`, type:`text`},
 		cartIds: {name:`cart`, type:`foreignKey-many`},
-		user_type: {name:`user_type`, type:`select`},
-		display_name: {name:`display_name`, type:`text`},
+		userType: {name:`user type`, type:`select`},
+		contactName: {name:`contact name`, type:`text`},
 		buyer_quotes: {name:`buyer_quotes`, type:`text`},
-		approval: {name:`approval`, type:`select`},
-		farmIds: {name:`farms`, type:`foreignKey-many`},
+		accountApproval: {name:`account approval`, type:`select`},
+		farmIdId: {name:`farm id`, type:`foreignKey-one`},
 		seller_quoteIds: {name:`seller_quotes`, type:`foreignKey-many`},
-		agencyIdIds: {name:`agency id`, type:`foreignKey-many`},
+		agencyIdId: {name:`agency id`, type:`foreignKey-one`},
+		checkoutProfileIds: {name:`checkout profiles`, type:`foreignKey-many`},
+		organization: {name:`organization`, type:`text`},
+		zipcode: {name:`zipcode`, type:`number`},
+		phone: {name:`phone`, type:`text`},
+		additionalComments: {name:`additional comments`, type:`multilineText`},
+		registrationApproval: {name:`registration approval`, type:`select`},
+		userId: {name:`user id`, type:`formula`},
 	},
 	"Farms": {
 		farmName: {name:`farm name`, type:`text`},
@@ -37,7 +49,7 @@ export const Columns = {
 		farmPicture: {name:`farm picture`, type:`multipleAttachment`},
 		address: {name:`address`, type:`text`},
 		zipCode: {name:`zip code`, type:`number`},
-		userIdIds: {name:`user id`, type:`foreignKey-many`},
+		userIdId: {name:`user id`, type:`foreignKey-one`},
 		market: {name:`market`, type:`multiSelect`},
 		pickUpOptions: {name:`pick up options`, type:`multiSelect`},
 		pacaperishableAgriculturalCommoditiesAct: {name:`PACA (Perishable Agricultural Commodities Act)`, type:`checkbox`},
@@ -45,73 +57,181 @@ export const Columns = {
 		farmingPracticeType: {name:`farming practice type`, type:`multiSelect`},
 		ableToDeliver: {name:`able to deliver`, type:`checkbox`},
 		coldChainCapabilities: {name:`cold chain capabilities`, type:`checkbox`},
-	},
-	"Clusters": {
-		clusterName: {name:`cluster name`, type:`text`},
-		refrigeration: {name:`refrigeration`, type:`checkbox`},
-		transportationOption: {name:`transportation option`, type:`select`},
-		isdraft: {name:`isDraft`, type:`checkbox`},
+		listingsCopyIds: {name:`Listings copy`, type:`foreignKey-many`},
+		reservedListingIds: {name:`Reserved Listings`, type:`foreignKey-many`},
+		additionalComments: {name:`additional comments`, type:`multilineText`},
+		website: {name:`website`, type:`text`},
+		contactName: {name:`contact name`, type:`text`},
+		addressLine1: {name:`address line 1`, type:`text`},
+		addressLine2: {name:`address line 2`, type:`text`},
+		additionalContactInformation: {name:`additional contact information`, type:`text`},
+		farmSize: {name:`farm size`, type:`select`},
+		approved: {name:`approved`, type:`checkbox`},
+		county: {name:`county`, type:`text`},
+		callDuring: {name:`call during`, type:`text`},
+		phoneString: {name:`phone string`, type:`text`},
+		zipcode: {name:`zipcode`, type:`number`},
+		legalName: {name:`legal name`, type:`text`},
+		farmNamePrivatized: {name:`farm name privatized`, type:`checkbox`},
+		addressPrivatized: {name:`address privatized`, type:`checkbox`},
+		farmPracticesPrivatized: {name:`farm practices privatized`, type:`checkbox`},
+		pacaPrivatized: {name:`paca privatized`, type:`checkbox`},
 	},
 	"Agencies": {
-		agencyName: {name:`agency name`, type:`text`},
-		ein: {name:`EIN`, type:`number`},
+		agencyId: {name:`agency id`, type:`formula`},
 		userIdId: {name:`user id`, type:`foreignKey-one`},
 		phone: {name:`phone`, type:`text`},
 		email: {name:`email`, type:`text`},
 		website: {name:`website`, type:`text`},
 		address: {name:`address`, type:`text`},
-		zipCode: {name:`zip code`, type:`number`},
+		zipcode: {name:`zipcode`, type:`number`},
 		socialMedia: {name:`social media`, type:`text`},
-		notes: {name:`notes`, type:`multilineText`},
-	},
-	"Listings": {
-		crop: {name:`crop`, type:`text`},
-		farmIdIds: {name:`farm id`, type:`foreignKey-many`},
-		description: {name:`description`, type:`multilineText`},
-		listingPicture: {name:`listing picture`, type:`multipleAttachment`},
-		expirationDate: {name:`expiration date`, type:`date`},
-		unitType: {name:`unit type`, type:`select`},
-		lbsPerUnit: {name:`lbs per unit`, type:`number`},
-		standardPricePerUnit: {name:`standard price per unit`, type:`number`},
-		restrictedPricePerUnit: {name:`restricted price per unit`, type:`number`},
-		dateEntered: {name:`date entered`, type:`date`},
-		availableUntil: {name:`available until`, type:`date`},
-		growingSeason: {name:`growing season`, type:`select`},
-		palletsAvailable: {name:`pallets available`, type:`number`},
-		palletsPending: {name:`pallets pending`, type:`number`},
-		palletsSold: {name:`pallets sold`, type:`number`},
-		unitsPerPallet: {name:`units per pallet`, type:`number`},
-		notes: {name:`notes`, type:`multilineText`},
-		distressed: {name:`distressed`, type:`checkbox`},
-		firstAvailableDate: {name:`first available date`, type:`date`},
-		reservedListingIds: {name:`reserved listings`, type:`foreignKey-many`},
-		privatized: {name:`privatized`, type:`checkbox`},
-		usersInterested: {name:`users interested`, type:`number`},
+		additionalComments: {name:`additional comments`, type:`multilineText`},
+		agencyName: {name:`agency name`, type:`text`},
+		populationServed: {name:`population served`, type:`multiSelect`},
+		contactName: {name:`contact name`, type:`text`},
+		legalName: {name:`legal name`, type:`text`},
+		addressLine1: {name:`address line 1`, type:`text`},
+		addressLine2: {name:`address line 2`, type:`text`},
+		federalTaxId: {name:`federal tax id`, type:`number`},
+		additionalContactInformation: {name:`additional contact information`, type:`text`},
 	},
 	"Quotes": {
 		quoteId: {name:`quote id`, type:`formula`},
-		totalCost: {name:`total cost`, type:`number`},
 		status: {name:`status`, type:`select`},
 		accepted: {name:`accepted`, type:`checkbox`},
 		buyerIdIds: {name:`buyer id`, type:`foreignKey-many`},
 		notificationIds: {name:`notifications`, type:`foreignKey-many`},
-		preCost: {name:`pre cost`, type:`number`},
+		initialCost: {name:`initial cost`, type:`number`},
 		reservedListingIds: {name:`reserved listings`, type:`foreignKey-many`},
 		createdat: {name:`createdAt`, type:`formula`},
 		resolved: {name:`resolved`, type:`checkbox`},
+		transportationCost: {name:`transportation cost`, type:`number`},
+		processingFees: {name:`processing fees`, type:`number`},
+		additionalTaxes: {name:`additional taxes`, type:`number`},
+		checkoutProfileId: {name:`checkout profile`, type:`foreignKey-one`},
+		sent: {name:`sent`, type:`checkbox`},
+		approved: {name:`approved`, type:`checkbox`},
+		processed: {name:`processed`, type:`checkbox`},
+		cancelled: {name:`cancelled`, type:`checkbox`},
+		farm2peopleDiscount: {name:`farm2people discount`, type:`number`},
+		discountReason: {name:`discount reason`, type:`multilineText`},
+		estimatedArrival: {name:`estimated arrival`, type:`date`},
 	},
-	"Reserved Listings": {
-		reservedListingId: {name:`reserved listing id`, type:`formula`},
-		pallets: {name:`pallets`, type:`number`},
-		userIdId: {name:`user id`, type:`foreignKey-one`},
-		listingIdIds: {name:`listing id`, type:`foreignKey-many`},
+	"Checkout Profiles": {
+		profileId: {name:`profile id`, type:`formula`},
+		firstName: {name:`first name`, type:`text`},
+		lastName: {name:`last name`, type:`text`},
+		phone: {name:`phone`, type:`text`},
+		email: {name:`email`, type:`text`},
+		addressLine1: {name:`address line 1`, type:`text`},
+		addressLine2: {name:`address line 2`, type:`text`},
+		city: {name:`city`, type:`text`},
+		state: {name:`state`, type:`text`},
+		country: {name:`country`, type:`text`},
+		zipcode: {name:`zipcode`, type:`text`},
 		quoteIdIds: {name:`quote id`, type:`foreignKey-many`},
-		resolved: {name:`resolved`, type:`checkbox`},
+		userIdId: {name:`user id`, type:`foreignKey-one`},
+		profileName: {name:`profile name`, type:`text`},
+		preferDelivery: {name:`prefer delivery`, type:`checkbox`},
+		canAch: {name:`can ACH`, type:`checkbox`},
+		canPaypal: {name:`can paypal`, type:`checkbox`},
+		canCard: {name:`can card`, type:`checkbox`},
+		canCheck: {name:`can check`, type:`checkbox`},
 	},
 	"Notifications": {
 		description: {name:`description`, type:`text`},
 		buyerId: {name:`buyer id`, type:`text`},
 		notificationType: {name:`notification type`, type:`select`},
 		quoteIds: {name:`quote`, type:`foreignKey-many`},
+	},
+	"Email": {
+		message: {name:`message`, type:`text`},
+		email: {name:`email`, type:`text`},
+		subject: {name:`subject`, type:`text`},
+		lastName: {name:`last name`, type:`text`},
+		phone: {name:`phone`, type:`text`},
+		firstName: {name:`first name`, type:`text`},
+	},
+	"Reserved Listings": {
+		reservedListingId: {name:`reserved listing id`, type:`formula`},
+		pallets: {name:`pallets`, type:`number`},
+		userIdId: {name:`user id`, type:`foreignKey-one`},
+		listingIdId: {name:`listing id`, type:`foreignKey-one`},
+		quoteIdId: {name:`quote id`, type:`foreignKey-one`},
+		resolved: {name:`resolved`, type:`checkbox`},
+		listingsCopyIds: {name:`Listings copy`, type:`foreignKey-many`},
+		farmIdId: {name:`farm id`, type:`foreignKey-one`},
+	},
+	"Listings": {
+		listingId: {name:`listing id`, type:`formula`},
+		farmIdId: {name:`farm id`, type:`foreignKey-one`},
+		details: {name:`details`, type:`multilineText`},
+		listingPicture: {name:`listing picture`, type:`multipleAttachment`},
+		sellByDate: {name:`sell by date`, type:`date`},
+		individualProduceUnit: {name:`individual produce unit`, type:`select`},
+		standardPricePerGroupedProduceType: {name:`standard price per grouped produce type`, type:`number`},
+		agencyPricePerGroupedProduceType: {name:`agency price per grouped produce type`, type:`number`},
+		dateEntered: {name:`date entered`, type:`date`},
+		availableUntil: {name:`available until`, type:`date`},
+		growingSeason: {name:`growing season`, type:`select`},
+		palletsAvailable: {name:`pallets available`, type:`number`},
+		palletsPending: {name:`pallets pending`, type:`number`},
+		palletsSold: {name:`pallets sold`, type:`number`},
+		groupedProduceTypePerPallet: {name:`grouped produce type per pallet`, type:`number`},
+		firstAvailableDate: {name:`first available date`, type:`date`},
+		reservedListingIds: {name:`reserved listings`, type:`foreignKey-many`},
+		privatized: {name:`privatized`, type:`checkbox`},
+		usersInterested: {name:`users interested`, type:`number`},
+		produceId: {name:`produce`, type:`foreignKey-one`},
+		groupedProduceType: {name:`grouped produce type`, type:`select`},
+		individualProduceUnitsPerGroupedProduceType: {name:`individual produce units per grouped produce type`, type:`number`},
+		groupedProduceTypePerMasterUnit: {name:`grouped produce type per master unit`, type:`number`},
+		palletsPerMasterPallet: {name:`pallets per master pallet`, type:`number`},
+		lbsPerGroupedProduceType: {name:`lbs per grouped produce type`, type:`number`},
+		hasMasterUnits: {name:`has master units`, type:`checkbox`},
+		hasMasterPallets: {name:`has master pallets`, type:`checkbox`},
+		masterUnitsPerPallet: {name:`master units per pallet`, type:`number`},
+		masterPallets: {name:`master pallets`, type:`number`},
+	},
+	"Listing UPDATED": {
+		listingId: {name:`listing id`, type:`formula`},
+		farmIdId: {name:`farm id`, type:`foreignKey-one`},
+		details: {name:`details`, type:`multilineText`},
+		listingPicture: {name:`listing picture`, type:`multipleAttachment`},
+		sellByDate: {name:`sell by date`, type:`date`},
+		individualProduceUnit: {name:`individual produce unit`, type:`select`},
+		standardPricePerGroupedProduceType: {name:`standard price per grouped produce type`, type:`number`},
+		agencyPricePerGroupedProduceType: {name:`agency price per grouped produce type`, type:`number`},
+		dateEntered: {name:`date entered`, type:`date`},
+		availableUntil: {name:`available until`, type:`date`},
+		growingSeason: {name:`growing season`, type:`select`},
+		palletsAvailable: {name:`pallets available`, type:`number`},
+		palletsPending: {name:`pallets pending`, type:`number`},
+		palletsSold: {name:`pallets sold`, type:`number`},
+		distressed: {name:`distressed`, type:`checkbox`},
+		firstAvailableDate: {name:`first available date`, type:`date`},
+		reservedListingIds: {name:`reserved listings`, type:`foreignKey-many`},
+		privatized: {name:`privatized`, type:`checkbox`},
+		usersInterested: {name:`users interested`, type:`number`},
+		produceId: {name:`produce`, type:`foreignKey-one`},
+		groupedProduceType: {name:`grouped produce type`, type:`select`},
+		individualProduceUnitsPerGroupedProduceType: {name:`individual produce units per grouped produce type`, type:`number`},
+		groupedProduceTypePerMasterUnit: {name:`grouped produce type per master unit`, type:`number`},
+		palletsPerMasterPallet: {name:`pallets per master pallet`, type:`number`},
+		lbsPerGroupedProduceType: {name:`lbs per grouped produce type`, type:`number`},
+		hasMasterUnits: {name:`has master units`, type:`checkbox`},
+		hasMasterPallets: {name:`has master pallets`, type:`checkbox`},
+		groupedProduceTypePerPallet: {name:`grouped produce type per pallet`, type:`number`},
+		masterUnitsPerPallet: {name:`master units per pallet`, type:`number`},
+		masterPallets: {name:`master pallets`, type:`number`},
+	},
+	"Produce Type": {
+		produceType: {name:`produce type`, type:`text`},
+		growingGroup: {name:`growing group`, type:`multiSelect`},
+		producePicture: {name:`produce picture`, type:`multipleAttachment`},
+		produceCategory: {name:`produce category`, type:`select`},
+		listingIdIds: {name:`listing ids`, type:`foreignKey-many`},
+		listingsCopyIds: {name:`Listings copy`, type:`foreignKey-many`},
 	},
 };
